pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "selenium>=4.35.0",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
]
